requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.3
djangorestframework==3.14.0
django-cors-headers